Simplify class service class hierarchy
