+++ src/clabe/apps/_bonsai.py
@@ -295,12 +295,28 @@ class BonsaiApp(App):
         return self
 
     def build_runner(self, allow_std_error: bool = False) -> Callable[[Launcher], Self]:
-        super_method = super().build_runner
+        """
+        Builds a runner function for the application.
+
+        This method returns a callable that can be executed by the launcher to run the application.
+
+        Args:
+            allow_std_error (bool): Whether to allow stderr in the output. Defaults to False.
+
+        Returns:
+            Callable[[Launcher], Self]: A callable that takes a launcher instance and returns the application instance.
+        """
 
         def _run(launcher: Launcher):
             """Internal wrapper function"""
-            self.add_app_settings(launcher=launcher)
-            return super_method(allow_std_error=allow_std_error)(launcher)
+            try:
+                self.add_app_settings(launcher=launcher)
+                self.run()
+                result = self.output_from_result(allow_stderr=allow_std_error)
+            except subprocess.CalledProcessError as e:
+                logger.error(f"App {self.__class__.__name__} failed with error: {e}")
+                raise
+            return result
 
         return _run
