+++ src/clabe/apps/_python_script.py
@@ -5,7 +5,7 @@ import os
 import shutil
 import subprocess
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Optional, Self
+from typing import TYPE_CHECKING, Any, Callable, Optional, Self
 
 from typing_extensions import override
 
@@ -113,6 +113,7 @@ class PythonScriptApp(App):
 
     @override
     def add_app_settings(self, **kwargs) -> Self:
+        """Adds settings to the application"""
         self._additional_arguments = " ".join([self._additional_arguments] + [f"--{k} {v}" for k, v in kwargs.items()])
         return self
 
@@ -292,3 +293,28 @@ class PythonScriptApp(App):
                 "uv is not installed in this computer. Please install uv. see https://docs.astral.sh/uv/getting-started/installation/"
             )
         return True
+
+    def build_runner(self, allow_std_error: bool = False) -> Callable[[Launcher], Self]:
+        """
+        Builds a runner function for the application.
+
+        This method returns a callable that can be executed by the launcher to run the application.
+
+        Args:
+            allow_std_error (bool): Whether to allow stderr in the output. Defaults to False.
+
+        Returns:
+            Callable[[Launcher], Self]: A callable that takes a launcher instance and returns the application instance.
+        """
+
+        def _run(launcher: Launcher):
+            """Internal wrapper function"""
+            try:
+                self.run()
+                result = self.output_from_result(allow_stderr=allow_std_error)
+            except subprocess.CalledProcessError as e:
+                logger.error(f"App {self.__class__.__name__} failed with error: {e}")
+                raise
+            return result
+
+        return _run
