+++ src/clabe/launcher/_base.py
@@ -130,7 +130,7 @@ class Launcher(Generic[TRig, TSession, TTaskLogic]):
             self.dispose()
 
         except KeyboardInterrupt:
-            logger.critical("User interrupted the process.")
+            logger.error("User interrupted the process.")
             self._exit(-1)
             return
 
@@ -601,13 +601,11 @@ class Launcher(Generic[TRig, TSession, TTaskLogic]):
                 if not self.settings.allow_dirty:
                     self.repository.try_prompt_full_reset(self.ui_helper, force_reset=False)
                     if self.repository.is_dirty_with_submodules():
-                        logger.critical(
-                            "Dirty repository not allowed. Exiting. Consider running with --allow-dirty flag."
-                        )
+                        logger.error("Dirty repository not allowed. Exiting. Consider running with --allow-dirty flag.")
                         self._exit(-1)
 
         except Exception as e:
-            logger.critical("Failed to validate dependencies. %s", e)
+            logger.error("Failed to validate dependencies. %s", e)
             self._exit(-1)
             raise e
 
@@ -637,7 +635,7 @@ class Launcher(Generic[TRig, TSession, TTaskLogic]):
             self.create_directory(self.temp_dir)
 
         except OSError as e:
-            logger.critical("Failed to create directory structure: %s", e)
+            logger.error("Failed to create directory structure: %s", e)
             self._exit(-1)
 
     @classmethod
